Remove allow(dead_code) lint
